services/auth/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
httpx==0.28.1
PyJWT==2.9.0
bcrypt==4.2.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis[asyncio]==5.2.1
prometheus-client
python-json-logger==2.0.7

pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
